requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.5.1+cu121
torchvision==0.20.1+cu121

numpy
scipy
pandas
scikit-learn
matplotlib
seaborn
pillow
tqdm
pyyaml

timm==1.0.12
captum==0.7.0
quantus==0.6.0
kaggle==1.6.17

